refactor: move violations to linters
eradicate, flake8_boolean_trap, flake8_commas,
flake8_import_conventions, flake8_no_pep420,
flake8_pie, isort
