+++ src/rules/flake8_import_conventions/rules.rs
@@ -1,9 +1,22 @@
+use ruff_macros::derive_message_formats;
 use rustc_hash::FxHashMap;
 use rustpython_ast::Stmt;
 
 use crate::ast::types::Range;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+
+define_violation!(
+    pub struct ImportAliasIsNotConventional(pub String, pub String);
+);
+impl Violation for ImportAliasIsNotConventional {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let ImportAliasIsNotConventional(name, asname) = self;
+        format!("`{name}` should be imported as `{asname}`")
+    }
+}
 
 /// ICN001
 pub fn check_conventional_import(
@@ -25,10 +38,7 @@ pub fn check_conventional_import(
         }
         if !is_valid_import {
             return Some(Diagnostic::new(
-                violations::ImportAliasIsNotConventional(
-                    name.to_string(),
-                    expected_alias.to_string(),
-                ),
+                ImportAliasIsNotConventional(name.to_string(), expected_alias.to_string()),
                 Range::from_located(import_from),
             ));
         }
