+++ src/rules/isort/rules/mod.rs
@@ -1,5 +1,5 @@
-pub use add_required_imports::add_required_imports;
-pub use organize_imports::organize_imports;
+pub use add_required_imports::{add_required_imports, MissingRequiredImport};
+pub use organize_imports::{organize_imports, UnsortedImports};
 
 pub mod add_required_imports;
 pub mod organize_imports;
